apps/web/backend/requirements.txt
Flask
Flask-CORS
librosa
numpy
scipy
pydub
python-dotenv
requests